docdiffops_mvp/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
celery==5.4.0
redis==5.2.1
PyYAML==6.0.2
pymupdf==1.25.2
rapidfuzz==3.11.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
lxml==5.3.0
